Use @rpath to help link .so's in Mac virtual environments.
This commit changes the way location information is encoded in a .so
file. Previously, the path would be hard-coded; now it's relative to
Python's library directory (get_python_lib()). By using @rpath, you
don't need to rebuild the project from source every time you install
it in a new location. However, clients of the project's .so's will
need to add ['-rpath', get_python_lib()] to their link arguments in
order to link to them properly. (This is demonstrated in setup.py.)